Namespace hygiene
  msghub, hub, hubmessage -> msghublib
  hubclient, hubconnection, msghub_impl, ihub -> msghublib::detail

  msghub::span is alias for std::span or msghublib::detail::span
  depending on compiler support
