+++ pub/msghub.h
@@ -1,5 +1,4 @@
-#ifndef _MSGHUB_MSGHUB_H_
-#define _MSGHUB_MSGHUB_H_
+#pragma once
 
 #include <string>
 #include <vector>
@@ -9,34 +8,36 @@
 #include <boost/asio.hpp>
 #include "span.h"
 
-class msghub_impl;
-class msghub
-{
-  public:
-    typedef std::function< void(std::string_view topic, span<char const> message) > onmessage;
-    
-  public:
-    explicit msghub(boost::asio::any_io_executor);
-    ~msghub();
+namespace msghublib {
 
-    bool connect(const std::string& hostip, uint16_t port);
-    bool create(uint16_t port);
+    namespace detail { class msghub_impl; }
 
-    bool unsubscribe(const std::string& topic);
-    bool subscribe(const std::string& topic, onmessage handler);
-    bool publish(std::string_view topic, span<char const> message);
+    class msghub
+    {
+      public:
+        typedef std::function< void(std::string_view topic, span<char const> message) > onmessage;
+        
+      public:
+        explicit msghub(boost::asio::any_io_executor);
+        ~msghub();
 
-    // Treat string literals specially, not including the terminating NUL
-    template <size_t N>
-	bool publish(std::string_view topic, char const (&literal)[N]) {
-        static_assert(N>0);
-        return publish(topic, span<char const>{literal, N-1});
-    }
+        bool connect(const std::string& hostip, uint16_t port);
+        bool create(uint16_t port);
 
-    void stop();
-    
-  private:
-    std::shared_ptr<msghub_impl> pimpl;
-};
+        bool unsubscribe(const std::string& topic);
+        bool subscribe(const std::string& topic, onmessage handler);
+        bool publish(std::string_view topic, span<char const> message);
 
-#endif
+        // Treat string literals specially, not including the terminating NUL
+        template <size_t N>
+        bool publish(std::string_view topic, char const (&literal)[N]) {
+            static_assert(N>0);
+            return publish(topic, span<char const>{literal, N-1});
+        }
+
+        void stop();
+        
+      private:
+        std::shared_ptr<detail::msghub_impl> pimpl;
+    };
+}
