+++ src/hubclient.h
@@ -1,26 +1,22 @@
-#ifndef _MSGHUB_HUBCLIENT_H_
-#define _MSGHUB_HUBCLIENT_H_
+#pragma once
 
-#include "hub.h"
-//#include "msghub.h"
+#include "ihub.h"
 #include "hubmessage.h"
 
 #include <memory>
-#include <cstdlib>
 #include <functional>
-#include <algorithm>
-#include <vector>
 #include <deque>
 
 #include <boost/asio.hpp>
 
+namespace msghublib { namespace detail {
 using boost::asio::ip::tcp;
 
 class hubclient : public std::enable_shared_from_this<hubclient>
 {
   public:
     template <typename Executor>
-	hubclient(Executor executor, hub& distrib)
+	hubclient(Executor executor, ihub& distrib)
       : socket_(make_strand(executor))
       , distributor_(distrib)
     {}
@@ -37,9 +33,9 @@ class hubclient : public std::enable_shared_from_this<hubclient>
     auto bind(void (hubclient::*)(error_code));
   private:
 	tcp::socket			socket_;
-	hub&				distributor_;
+	ihub&				distributor_;
 	hubmessage			inmsg_;
 	hubmessage_queue	outmsg_queue_;
 };
 
-#endif
+} }
