+++ src/msghub.cpp
@@ -4,37 +4,41 @@
 
 #include <boost/asio.hpp>
 
-msghub::~msghub() = default;
+namespace msghublib {
 
-/*explicit*/ msghub::msghub(boost::asio::any_io_executor executor)
- : pimpl(std::make_shared<msghub_impl>(executor))
-{ }
+    msghub::~msghub() = default;
 
-void msghub::stop() {
-	return pimpl->stop();
-}
+    /*explicit*/ msghub::msghub(boost::asio::any_io_executor executor)
+     : pimpl(std::make_shared<detail::msghub_impl>(executor))
+    { }
 
-bool msghub::connect(const std::string& hostip, uint16_t port)
-{
-	return pimpl->connect(hostip, port);
-}
+    void msghub::stop() {
+        return pimpl->stop();
+    }
 
-bool msghub::create(uint16_t port)
-{
-	return pimpl->create(port);
-}
+    bool msghub::connect(const std::string& hostip, uint16_t port)
+    {
+        return pimpl->connect(hostip, port);
+    }
 
-bool msghub::unsubscribe(const std::string& topic)
-{
-	return pimpl->unsubscribe(topic);
-}
+    bool msghub::create(uint16_t port)
+    {
+        return pimpl->create(port);
+    }
 
-bool msghub::subscribe(const std::string& topic, onmessage handler)
-{
-	return pimpl->subscribe(topic, handler);
-}
+    bool msghub::unsubscribe(const std::string& topic)
+    {
+        return pimpl->unsubscribe(topic);
+    }
 
-bool msghub::publish(std::string_view topic, span<char const> message)
-{
-	return pimpl->publish(topic, message);
+    bool msghub::subscribe(const std::string& topic, onmessage handler)
+    {
+        return pimpl->subscribe(topic, handler);
+    }
+
+    bool msghub::publish(std::string_view topic, span<char const> message)
+    {
+        return pimpl->publish(topic, message);
+    }
+
 }
