+++ src/msghub_impl.h
@@ -1,7 +1,6 @@
-#ifndef _MSGHUB_MSGHUB_IMPL_H_
-#define _MSGHUB_MSGHUB_IMPL_H_
+#pragma once
 
-#include "hub.h"
+#include "ihub.h"
 #include "msghub.h"
 #include "hubconnection.h"
 #include "hubclient.h"
@@ -16,40 +15,42 @@
 
 using boost::asio::ip::tcp;
 
-class msghub_impl
-  : public hub,
-    public std::enable_shared_from_this<msghub_impl>
-{
-  public:
-    using any_io_executor = boost::asio::any_io_executor;
+namespace msghublib { namespace detail {
 
-  private:
-	std::map<std::string, msghub::onmessage> messagemap_;
+    class msghub_impl
+      : public ihub,
+        public std::enable_shared_from_this<msghub_impl>
+    {
+      public:
+        using any_io_executor = boost::asio::any_io_executor;
 
-  private:
-	tcp::acceptor acceptor_;
-    boost::asio::executor_work_guard<any_io_executor> work_;
-	std::shared_ptr<hubconnection> publisher_;
+      private:
+        std::map<std::string, msghub::onmessage> messagemap_;
 
-  public:
-	msghub_impl(any_io_executor io_service);
-	~msghub_impl();
-	bool connect(const std::string& hostip, uint16_t port);
-	bool create(uint16_t port);
-	bool publish(std::string_view topic, span<char const> message);
+      private:
+        tcp::acceptor acceptor_;
+        boost::asio::executor_work_guard<any_io_executor> work_;
+        std::shared_ptr<hubconnection> publisher_;
 
-	bool unsubscribe(const std::string& topic);
-	bool subscribe(const std::string& topic, msghub::onmessage handler);
+      public:
+        msghub_impl(any_io_executor io_service);
+        ~msghub_impl();
+        bool connect(const std::string& hostip, uint16_t port);
+        bool create(uint16_t port);
+        bool publish(std::string_view topic, span<char const> message);
 
-    void stop();
+        bool unsubscribe(const std::string& topic);
+        bool subscribe(const std::string& topic, msghub::onmessage handler);
 
-  public:
-	std::multimap<std::string, std::shared_ptr<hubclient> > client_subs_;
+        void stop();
 
-	void distribute(std::shared_ptr<hubclient> const& subscriber, hubmessage const& msg);
-	void deliver(hubmessage const& msg);
-	void accept_next();
-	void handle_accept(std::shared_ptr<hubclient> const& session, const boost::system::error_code& error);
-};
+      public:
+        std::multimap<std::string, std::shared_ptr<hubclient> > client_subs_;
 
-#endif
+        void distribute(std::shared_ptr<hubclient> const& subscriber, hubmessage const& msg);
+        void deliver(hubmessage const& msg);
+        void accept_next();
+        void handle_accept(std::shared_ptr<hubclient> const& session, const boost::system::error_code& error);
+    };
+
+} }
